Added more sanity checking for amplicon BED file creation

+++ code/runApp.sh
@@ -1,13 +1,16 @@
 #!/bin/sh
 
+LIM=""
+LIM="-#25k"
+
 code=`dirname $0`
 usage="$0  <sampleID> <r1.fq> <r2.fq> <hg19.fasta> <manifest.txt>"
 if [ $# -lt 5 ];then
 	echo $usage
 	echo "Code repo is $code"
+	echo "Debug:Lim=$LIM,Host=`hostname`,Version=`novoalign version`"
 	exit
 fi
-
 if [ ! -d $code ];then
 	echo No code repo found for scripts
 	echo exiting
@@ -33,7 +36,10 @@ amplicon_noprimer="$outdir/manifest2bed/ampliconNoPrimerRegion.bed"
 variants="$outdir/variants.vcf.gz"
 bam="$outdir/alignments.bam"
 
+
+novoversion=`novoalign version`
 mkdir -p $outdir $variantdir
+echo "Novoalign $novoversion"
 echo `date` Pipeline started on `hostname`
 echo "Input: SID: $sid, Genome=$REFGENOME, fastqs=$fastqs, manifest=$manifest"
 
@@ -48,11 +54,19 @@ if [ ! -e $ampliconbed ];then
 perl $code/manifestToAmplicon.pl $manifest -outdir $outdir/manifest2bed/
 fi
 
-#Alignment
+#check for existence of amplicon BED file
+if [ ! -e $ampliconbed ];then
+	echo "Error: No Amplicon bed file found in $ampliconbed"
+	echo "exiting"
+	exit 
+fi
+
+#Alignment and Bam-sorting
 if [ ! -e $bam ];then
-	echo `date` Align and Sorting
-	novoalign -#40k -d $GENOMEIDX -f $fastqs --amplicons $ampliconbed -oSAM -rR -k  2> $outdir/novoalign_log.txt |  samtools view -uS - > $outdir/novoalign.raw.bam
-	novosort $outdir/novoalign.raw.bam -o $bam -i  2>log.txt 
+	echo `date` Aliging reads with Novoalign $novoversion on `hostname`
+	novoalign $LIM -d $GENOMEIDX -f $fastqs --amplicons $ampliconbed -oSAM -rR -k  2> $outdir/novoalign_log.txt |  samtools view -uS - > $outdir/novoalign.raw.bam
+	echo `date` Sorting alignments 
+	novosort $outdir/novoalign.raw.bam -o $bam -i  2>$outdir/novosort_log.txt 
 	rm  $outdir/novoalign.raw.bam
 fi
 
@@ -64,7 +78,8 @@ rm $GENOMEIDX
 if [ ! -e "$outdir/novo_coverage.tsv" ];then
 echo `date` Calculating Coverage
 namecol=16
-intersectBed -abam $bam  -b  $amplicon_noprimer  -f 0.99999 -bed -wa -wb |sort -k $namecol,$namecol | groupBy -g $namecol -c 1 -o count > $outdir/novo_coverage.tsv
+echo -e "TargetAmplicon\tCount" > $outdir/novo_coverage.tsv
+intersectBed -abam $bam  -b  $amplicon_noprimer  -f 0.99999 -bed -wa -wb |sort -k $namecol,$namecol | groupBy -g $namecol -c 1 -o count >> $outdir/novo_coverage.tsv
 #Plot graph
 #Rscript coverage.rscript
 fi
